category, address, cart, forms middleware added

+++ rest/routes/AddressRoutes.php
@@ -10,7 +10,12 @@ use Firebase\JWT\Key;
  * )
  */
 Flight::route("GET /addresses", function(){
-    Flight::json(Flight::address_service()->get_all());
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        Flight::json(Flight::address_service()->get_all());
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 
 /**
@@ -20,7 +25,12 @@ Flight::route("GET /addresses", function(){
   * )
   */
 Flight::route("GET /addresses/@id", function($id){
-    Flight::json(Flight::address_service()->get_by_id($id));
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        Flight::json(Flight::address_service()->get_by_id($id));
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 
 
@@ -51,14 +61,19 @@ Flight::route("GET /addresses/@id", function($id){
 * )
 */
 Flight::route("POST /addresses", function(){
-    $request = Flight::request()->data->getData(); 
-    $user_id = $request['user_id'];
-    $alias = $request['alias'];
-    $existing_alias = Flight::address_service()->get_address_by_user_id_and_alias($user_id, $alias);  
-    if(count($existing_alias) > 0){
-        Flight::json(["message" => "You already have address with that alias. Please choose another alias"], 404);
-    } else { 
-        Flight::json(['message' => "address added successfully", 'data' => Flight::address_service()->add($request)]);
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        $request = Flight::request()->data->getData(); 
+        $user_id = $request['user_id'];
+        $alias = $request['alias'];
+        $existing_alias = Flight::address_service()->get_address_by_user_id_and_alias($user_id, $alias);  
+        if(count($existing_alias) > 0){
+            Flight::json(["message" => "You already have address with that alias. Please choose another alias"], 404);
+        } else { 
+            Flight::json(['message' => "address added successfully", 'data' => Flight::address_service()->add($request)]);
+        }
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
     }
 });
 
@@ -91,15 +106,20 @@ Flight::route("POST /addresses", function(){
  * )
  */
 Flight::route("PUT /addresses/@id", function($id){
-    $address = Flight::request()->data->getData();
-    $user_id = $address['user_id'];
-    $alias = $address['alias'];
-    $existing_alias = Flight::address_service()->get_address_by_user_id_and_alias($user_id, $alias);
-    if(count($existing_alias) > 0 && $existing_alias[0]['id'] != $id){
-        Flight::json(["message" => "You already have address with that alias. Please choose another alias"], 404);
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        $address = Flight::request()->data->getData();
+        $user_id = $address['user_id'];
+        $alias = $address['alias'];
+        $existing_alias = Flight::address_service()->get_address_by_user_id_and_alias($user_id, $alias);
+        if(count($existing_alias) > 0 && $existing_alias[0]['id'] != $id){
+            Flight::json(["message" => "You already have address with that alias. Please choose another alias"], 404);
+        } else {
+            Flight::json(['message' => "address edited successfully", 'data' => Flight::address_service()->update($address, $id)]);
+        }
     } else {
-        Flight::json(['message' => "address edited successfully", 'data' => Flight::address_service()->update($address, $id)]);
-    }    
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }   
     
 });
 
@@ -121,8 +141,13 @@ Flight::route("PUT /addresses/@id", function($id){
  * )
  */
 Flight::route("DELETE /addresses/@id", function($id){
-    Flight::address_service()->delete($id);
-    Flight::json(['message' => "address deleted successfully"]);
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        Flight::address_service()->delete($id);
+        Flight::json(['message' => "address deleted successfully"]);
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 
 
@@ -133,7 +158,12 @@ Flight::route("DELETE /addresses/@id", function($id){
   * )
   */
 Flight::route("GET /addresses_by_user_id/@user_id", function($user_id){
-    Flight::json(Flight::address_service()->get_address_by_user_id($user_id));
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        Flight::json(Flight::address_service()->get_address_by_user_id($user_id));
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 
 
@@ -144,8 +174,13 @@ Flight::route("GET /addresses_by_user_id/@user_id", function($user_id){
   * )
   */
 Flight::route("GET /addresses_by_user_token/@user_id", function($user_id){
-    $decoded = (array)JWT::decode($user_id, new Key(Config::JWT_SECRET(), 'HS256'));
-    $decoded_user_id = $decoded['id'];
-    Flight::json(Flight::address_service()->get_address_by_user_id($decoded_user_id));
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        $decoded = (array)JWT::decode($user_id, new Key(Config::JWT_SECRET(), 'HS256'));
+        $decoded_user_id = $decoded['id'];
+        Flight::json(Flight::address_service()->get_address_by_user_id($decoded_user_id));
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 ?>
