+++ rest/routes/CartRoutes.php
@@ -10,7 +10,12 @@ use Firebase\JWT\Key;
  * )
  */
 Flight::route("GET /carts", function(){
-    Flight::json(Flight::cart_service()->get_all());
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        Flight::json(Flight::cart_service()->get_all());
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 
 
@@ -21,7 +26,12 @@ Flight::route("GET /carts", function(){
   * )
   */
 Flight::route("GET /carts/@id", function($id){
-    Flight::json(Flight::cart_service()->get_by_id($id));
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        Flight::json(Flight::cart_service()->get_by_id($id));
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 
 
@@ -49,12 +59,17 @@ Flight::route("GET /carts/@id", function($id){
 * )
 */
 Flight::route("POST /carts", function(){
-    $request = Flight::request()->data->getData(); 
-    $user_id = $request['user_id'];
-    $decoded = (array)JWT::decode($user_id, new Key(Config::JWT_SECRET(), 'HS256'));
-    $decoded_user_id = $decoded['id'];  
-    $request['user_id'] = $decoded_user_id;
-    Flight::json(['message' => "cart added successfully", 'data' => Flight::cart_service()->add($request)]);
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        $request = Flight::request()->data->getData(); 
+        $user_id = $request['user_id'];
+        $decoded = (array)JWT::decode($user_id, new Key(Config::JWT_SECRET(), 'HS256'));
+        $decoded_user_id = $decoded['id'];  
+        $request['user_id'] = $decoded_user_id;
+        Flight::json(['message' => "cart added successfully", 'data' => Flight::cart_service()->add($request)]);
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 
 
@@ -83,8 +98,13 @@ Flight::route("POST /carts", function(){
  * )
  */
 Flight::route("PUT /carts/@id", function($id){
-    $address = Flight::request()->data->getData();    
-    Flight::json(['message' => "cart edited successfully", 'data' => Flight::cart_service()->update($address, $id)]);
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        $address = Flight::request()->data->getData();    
+        Flight::json(['message' => "cart edited successfully", 'data' => Flight::cart_service()->update($address, $id)]);
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 
 
@@ -105,8 +125,13 @@ Flight::route("PUT /carts/@id", function($id){
  * )
  */
 Flight::route("DELETE /carts/@id", function($id){
-    Flight::cart_service()->delete($id);
-    Flight::json(['message' => "cart deleted successfully"]);
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        Flight::cart_service()->delete($id);
+        Flight::json(['message' => "cart deleted successfully"]);
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 
 /**
@@ -116,9 +141,14 @@ Flight::route("DELETE /carts/@id", function($id){
   * )
   */
 Flight::route("GET /carts_by_user_id/@user_id", function($user_id){
-    $decoded = (array)JWT::decode($user_id, new Key(Config::JWT_SECRET(), 'HS256'));
-    $decoded_user_id = $decoded['id'];
-    Flight::json(Flight::cart_service()->get_by_user_id($decoded_user_id));
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        $decoded = (array)JWT::decode($user_id, new Key(Config::JWT_SECRET(), 'HS256'));
+        $decoded_user_id = $decoded['id'];
+        Flight::json(Flight::cart_service()->get_by_user_id($decoded_user_id));
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 
 /**
@@ -138,10 +168,15 @@ Flight::route("GET /carts_by_user_id/@user_id", function($user_id){
  * )
  */
 Flight::route("DELETE /carts_by_user_id/@user_id", function($user_id){
-    $decoded = (array)JWT::decode($user_id, new Key(Config::JWT_SECRET(), 'HS256'));
-    $decoded_user_id = $decoded['id'];
-    Flight::cart_service()->delete_by_user_id($decoded_user_id);
-    Flight::json(['message' => "cart deleted successfully"]);
+    $user = Flight::get('user');
+    if(isset($user) && $user['authorization'] == 'unauthorized') {
+        $decoded = (array)JWT::decode($user_id, new Key(Config::JWT_SECRET(), 'HS256'));
+        $decoded_user_id = $decoded['id'];
+        Flight::cart_service()->delete_by_user_id($decoded_user_id);
+        Flight::json(['message' => "cart deleted successfully"]);
+    } else {
+        Flight::json(["message" => "User token doesn't exist!"], 404);
+    }
 });
 
 ?>
